test_task/requirements.txt
asgiref==3.7.2
certifi==2023.11.17      
charset-normalizer==3.3.2
dj-database-url==2.1.0   
Django==5.0              
idna==3.6
python-decouple==3.8
requests==2.31.0
sqlparse==0.4.4
stripe==7.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
